changed Decimal to float

+++ src/pdfparser/sqlmodel.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from sqlmodel import Field, SQLModel, Relationship
 
 class Document(SQLModel, table=True):
@@ -21,9 +20,9 @@ class Page(SQLModel, table=True):
 class BoundingBox(SQLModel, table=True):
         id: int | None = Field(default=None, primary_key=True)
         word_text: str
-        x: Decimal
-        y: Decimal
-        width: Decimal
-        height: Decimal
+        x: float
+        y: float
+        width: float
+        height: float
         page_id: int | None = Field(default=None, foreign_key="page.id")
         page: Page | None = Relationship(back_populates="bounding_boxes")
